[PIDM-1928] minor refactor
