+++ src/main/java/it/gov/pagopa/pdf/engine/util/Constants.java
@@ -4,6 +4,8 @@ public class Constants {
 
     public static final String UNZIPPED_FILES_FOLDER = "/unzipped";
     public static final String ZIP_FILE_NAME = "/input.zip";
+    public static final String DATA_KEY = "data";
+    public static final String TEMPLATE_KEY = "template";
 
     private Constants() {}
 }
