+++ src/test/java/it/gov/pagopa/pdf/engine/service/impl/GeneratePDFServiceImplTest.java
@@ -121,7 +121,7 @@ class GeneratePDFServiceImplTest {
     private static <T> void setMock(Class<T> classToMock, T mock) {
         try {
             Class<?> holder = Class.forName(classToMock.getName() + "$Holder");
-            Field instance = holder.getDeclaredField("INSTANCE");
+            Field instance = holder.getDeclaredField("instance");
             instance.setAccessible(true);
             instance.set(null, mock);
         } catch (Exception e) {
